+++ system/src/Grav/Framework/File/Formatter/AbstractFormatter.php
@@ -49,6 +49,16 @@ abstract class AbstractFormatter implements FileFormatterInterface
     }
 
     /**
+     * @return string
+     */
+    public function getMimeType(): string
+    {
+        $mime = $this->getConfig('mime');
+
+        return \is_string($mime) ? $mime : 'application/octet-stream';
+    }
+
+    /**
      * {@inheritdoc}
      * @see FileFormatterInterface::getDefaultFileExtension()
      */